Response class changes

+++ app/src/main/java/com/selvaraj/forecastweatherapp/model/response/Main.kt
@@ -38,5 +38,5 @@ data class Main(
 
     @SerializedName("temp_kf")
     @Expose
-    val tempKf: Int? = null
+    val tempKf: Double? = null
 )
